Use modern locations for configuration files
Back in the day, the way to do this was just to put your config files
willy-nilly wherever you want. This lead usually to a .foo file or
directory containing foo's configuration, and eventually, to a very
cluttered, disorganised home directory.

The modern solution is to either respect XDG_CONFIG_HOME and create a
directory foo in there, or to default to .config/foo. With this patch,
nice editor does exactly that.